assets: split code into modules and export at top-level
